embedder/pyproject.toml
[project]
name = "embedder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ollama>=0.5.1",
    "pydantic>=2.11.7",
    "pytest>=8.4.1",
    "quixstreams>=3.17.0",
    "weaviate-client>=4.15.3",
]
